requirements.txt
fastapi==0.103.2
fastapi==0.115.5
SQLAlchemy==2.0.0
Alembic==1.10.0
bcrypt==4.1.3
PyJWT==2.8.0
PyMySQL==1.1.0
pyodbc==5.1.0
pytest==7.2.2
pytest-cov==4.0.0
gunicorn==22.0.0
SQLAlchemy==2.0.0
openai==0.27.4
uvicorn==0.32.1
python-dotenv==1.0.0
openai==1.55.3
pydantic==2.9.2
tenacity==8.2.3